+++ noodles-bgzf/src/io.rs
@@ -4,5 +4,6 @@ mod buf_read;
 mod read;
 pub mod reader;
 mod seek;
+pub mod writer;
 
-pub use self::{buf_read::BufRead, read::Read, reader::Reader, seek::Seek};
+pub use self::{buf_read::BufRead, read::Read, reader::Reader, seek::Seek, writer::Writer};